requirements.txt
aiohttp>=3.9.0
pyyaml>=6.0
pandas>=2.2.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
tqdm>=4.66.0
matplotlib>=3.8.0
requests>=2.31.0
streamlit>=1.32.0
plotly>=5.19.0
numpy>=1.26.0
